generator-springboot/src/main/java/com/dbr/generator/springboot/system/mail/service/MailService.java
package com.dbr.generator.springboot.system.mail.service;

import lombok.RequiredArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.mail.javamail.MimeMessageHelper;
import org.springframework.stereotype.Service;
import org.springframework.web.multipart.MultipartFile;

import javax.mail.MessagingException;
import javax.mail.internet.MimeMessage;

@Service
@RequiredArgsConstructor
public class MailService {

    @Autowired
    private JavaMailSender javaMailSender;

    public void sendMail(String to, String subject, String text, boolean html, MultipartFile attachmentFile)
            throws MessagingException {
        MimeMessage msg = javaMailSender.createMimeMessage();
        MimeMessageHelper helper = new MimeMessageHelper(msg, true);
        helper.setTo(to);
        helper.setSubject(subject);
        helper.setText(text, html);
        if (attachmentFile != null) {
            helper.addAttachment(attachmentFile.getOriginalFilename(), attachmentFile);
        }
        javaMailSender.send(msg);
    }

}
